+++ src/main/java/br/com/grupomult/repositories/CarroRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import br.com.grupomult.entities.Carro;
 
-public interface CarroRepository extends CrudRepository<Carro, Integer>{
+public interface CarroRepository extends CrudRepository<Carro, Long>{
 
 }
